ozone_v4/requirements.txt
Django==4.2.8
django_environ==0.11.2
django_import_export==3.3.7
django_modeltranslation==0.18.11
djangorestframework==3.14.0
django-cleanup==8.0.0
gunicorn==21.2.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4